fixed parameter type for torch.fake_quantize per channel

+++ torch2trt/contrib/qat/layers/quant_pooling.py
@@ -98,8 +98,8 @@ class QuantMaxPool2d(torch.nn.Module,_utils.QuantInputMixin):
         else:
             quant_input = torch.fake_quantize_per_channel_affine(input,
                     quantizer.quant_scale,
-                    quantizer.zero_point.to(torch.long),
-                    quantizer.quant_axis.to(torch.long).item(),
+                    quantizer.zero_point,
+                    quantizer.quant_axis.to(torch.int32).item(),
                     quantizer.quant_min.to(torch.long).item(),
                     quantizer.quant_max.to(torch.long).item())
 
@@ -197,8 +197,8 @@ class QuantAdaptiveAvgPool2d(torch.nn.Module,_utils.QuantInputMixin):
         else:
             quant_input = torch.fake_quantize_per_channel_affine(input,
                     quantizer.quant_scale,
-                    quantizer.zero_point.to(torch.long),
-                    quantizer.quant_axis.to(torch.long).item(),
+                    quantizer.zero_point,
+                    quantizer.quant_axis.to(torch.int32).item(),
                     quantizer.quant_min.to(torch.long).item(),
                     quantizer.quant_max.to(torch.long).item())
